fix(mts): Moved common enterprise MTs endpoints to correct directory
