+++ src/CrowdinApiClient/Api/Enterprise/MachineTranslationEngineApi.php
@@ -2,9 +2,8 @@
 
 namespace CrowdinApiClient\Api\Enterprise;
 
-use CrowdinApiClient\Api\AbstractApi;
+use CrowdinApiClient\Api\MachineTranslationEngineApi as CrowdinMachineTranslationEngineApi;
 use CrowdinApiClient\Model\MachineTranslationEngine;
-use CrowdinApiClient\ModelCollection;
 
 /**
  * Machine Translation Engines (MTE) are the sources for pre-translations.
@@ -13,35 +12,8 @@ use CrowdinApiClient\ModelCollection;
  *
  * @package Crowdin\Api\Enterprise
  */
-class MachineTranslationEngineApi extends AbstractApi
+class MachineTranslationEngineApi extends CrowdinMachineTranslationEngineApi
 {
-    /**
-     * List MTs
-     * @link https://developer.crowdin.com/enterprise/api/v2/#operation/api.mts.getMany API Documentation
-     *
-     * @param array $params
-     * integer $params[groupId]<br>
-     * integer $params[limit]<br>
-     * integer $params[offset]
-     * @return ModelCollection
-     */
-    public function list(array $params = []): ModelCollection
-    {
-        return $this->_list('mts', MachineTranslationEngine::class, $params);
-    }
-
-    /**
-     * Get MT
-     * @link https://developer.crowdin.com/enterprise/api/v2/#operation/api.mts.get API Documentation
-     *
-     * @param int $mtId
-     * @return MachineTranslationEngine|null
-     */
-    public function get(int $mtId): ?MachineTranslationEngine
-    {
-        return $this->_get('mts/' . $mtId, MachineTranslationEngine::class);
-    }
-
     /**
      * Add MT
      * @link https://developer.crowdin.com/enterprise/api/v2/#operation/api.mts.post API Documentation
